requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
llama-cpp-python
httpx
numpy
sentence-transformers
python-multipart
aiofiles
huggingface-hub
